requirements.txt
python-telegram-bot==20.7
pandas==2.1.4
pandas-ta==0.3.14b0
mplfinance==0.12.10a0
matplotlib==3.7.4
numpy==1.25.2
SQLAlchemy==2.0.23
requests==2.31.0
websocket-client==1.6.4
pytz==2023.3
python-dotenv==1.0.0
Flask==2.3.3
